refactor: extract task date calculation service
